% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiCAF_plot.R
\name{Aqua_CAF}
\alias{Aqua_CAF}
\title{Graphing Cumulative Age Frequencies for Multiple Species}
\usage{
Aqua_CAF(df, age, CAF, species)
}
\arguments{
\item{df}{A dataframe with columns: species, age, and cumulative age frequency values for multiple fish species}

\item{age}{The age class of fish by year}

\item{CAF}{Cumulative age frequency values}

\item{species}{A species identifier for each row with age class and cumulative age frequency value}
}
\value{
Produces cumulative age frequency line plots for species contained in the dataframe
}
\description{
Graphing Cumulative Age Frequencies for Multiple Species
}
\examples{
Aqua_CAF(dataframe, age, CAF, species)
}
