% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wild_Citations.R
\name{wild_citations}
\alias{wild_citations}
\title{Publications used for wild population age structures}
\usage{
wild_citations()
}
\arguments{
\item{WildCITE}{Dataset; Contains species common names and full citation for age structure data used in AquaSpan package}

\item{species}{Fish species common name}

\item{citation}{Full citation information for the age structure data used for a particular species by the AquaSpan package}
}
\value{
The citation for the publication used to produce age structure plots
}
\description{
Publications used for wild population age structures
}
\examples{
wild_citations()
}
