% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WildDATA}
\alias{WildDATA}
\title{Wild Fish Age Structure Data}
\format{
A data frame with 206 rows and 5 variables
}
\source{
{Compiled by M. O'Neill from the following publications:

Blackfin Snapper (L. buccanella): Burton, M.L., Potts, J.C. & Carr, D.R. (2016). Age, growth, and natural mortality of blackfin snapper, Lutjanus buccanella, from the southeastern United States and U.S. Caribbean. Gulf and Caribbean Research, 27(1), 66-73. doi:10.18785/gcr.2701.10

Blue Runner (C. crysos): Goodwin, J.M. IV & Johnson, A.G. (1986). Age, growth, and mortality of blue runner, Caranx crysos, from the northern Gulf of Mexico. Northeast Gulf Science, 8(2), 107-114. doi:10.18785/negs.0802.02

Bluestriped Grunt (H. sciurus): Pitt, J.M., Trott, T.M. & Luckhurst, B.E. (2010). Bluestriped grunt (Haemulon sciurus) in Bermuda: age, growth, and reproduction studies. \link{Conference Paper}. 62nd Gulf and Caribbean Fisheries Institute, Cumana, Venezuela

French Angelfish (P. paru): Feitosa CV, Araújo ME, and Ferreira BP. Estimates on age, growth and mortality of the French angelfish Pomacanthus paru (Bloch, 1787) (Teleostei: Pomacanthidae) in the southwestern Atlantic. J Appl Ichthyol. 2017;00:1–6. doi:10.1111/jai.13246.

Jolthead Porgy

Lane Snapper

Permit (T. falcatus): Crabtree, R. E., Hood, P. B., and Snodgrass D. (2002) Age, growth, and reproduction of permit (Trachinotus falcatus) in Florida water. Fish. Bull. 100:26-34

Queen Triggerfish (B. vetula): Albuquerque, C.Q., Martins, A.S., Leite, N.O. Jr., Araujo, J.N, & Ribeiro, A.M. (2011). Age and growth of the queen triggerfish Balistes vetula (Tetraodontiformes, Balistidae) of the central coast of Brazil. Brazilian Journal of Oceanography, 59(3), 231-239

Snowy Grouper (H. niveatus): Kowal, K. (2010). Aspects of the life history of the snowy grouper, Epinephelus niveatus, in the Gulf of Mexico. Graduate Theses and Dissertations, http://scholarcommons.usf.edu/etd/3505

Tomtate (H. aurolineatum): Manooch, C.S. III & Barans, C.A. (1982). Distribution, abundance, and age and growth of the tomtate, Haemulon aurolineatum, along the southeastern United States coast. Fishery Bulletin, 80(1), 1-19

White Grunt (H.plumierii): Araujo, J. N. and A. S. Martins. (2007). Age, growth and mortality of white grunt (Haemulon plumierii) from the central coast of Brazil. Scientia Marina, 71(4). 793-800

}
}
\usage{
WildDATA
}
\description{
Contains cumulative age frequencies and individual counts per age class (in years), for several subtropical and tropical Western-Atlantic species of fish.
}
\examples{
data(WildDATA)
}
\keyword{datasets}
